+++ packages/shared/src/catalog/index.ts
@@ -4,6 +4,7 @@ export * from "./outputs/inspectorFields";
 export * from "./primitives/breadcrumb";
 export * from "./primitives/breadcrumbs";
 export * from "./primitives/button";
+export * from "./primitives/checkbox";
 export * from "./primitives/colorField";
 export * from "./primitives/dateField";
 export * from "./primitives/fileTrigger";